renaming/restructuring files for package
